Moved markdown vue files to their own folder
